Raise coverage to 94.5% and fix discover request ID

+++ internal/explorer/info.go
@@ -59,7 +59,7 @@ func rawDiscoverSupportedVersions(ctx context.Context, url, requestedVersion str
 	if err != nil {
 		return nil, err
 	}
-	id, err := jsonrpc.MakeID(1)
+	id, err := jsonrpc.MakeID(float64(1))
 	if err != nil {
 		return nil, err
 	}
